GDEV056 added switchable URL for related environment and dependencies update 🙄
