+++ package.json
@@ -3,11 +3,11 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@fortawesome/fontawesome-svg-core": "^1.2.26",
-    "@fortawesome/free-solid-svg-icons": "^5.12.0",
+    "@fortawesome/fontawesome-svg-core": "^1.2.27",
+    "@fortawesome/free-solid-svg-icons": "^5.12.1",
     "@fortawesome/react-fontawesome": "^0.1.8",
     "animate.css": "^3.7.2",
-    "axios": "^0.19.0",
+    "axios": "^0.19.2",
     "axos": "^0.0.1",
     "history": "^4.10.1",
     "jquery": "^3.4.1",
@@ -19,17 +19,17 @@
     "react-dom": "^16.12.0",
     "react-lightgallery": "^0.6.3",
     "react-moment": "^0.9.7",
-    "react-redux": "^7.1.3",
+    "react-redux": "^7.2.0",
     "react-router": "^5.1.2",
     "react-router-dom": "^5.1.2",
-    "react-scripts": "3.3.0",
+    "react-scripts": "3.4.0",
     "react-select": "^3.0.8",
-    "react-slider": "^1.0.1",
+    "react-slider": "^1.0.3",
     "react-toastify": "^5.5.0",
     "redux": "^4.0.5",
     "redux-thunk": "^2.3.0",
-    "styled-components": "^4.4.1",
-    "swiper": "^5.2.1"
+    "styled-components": "^5.0.1",
+    "swiper": "^5.3.1"
   },
   "scripts": {
     "start": "react-scripts start",
